+++ MudBlazor.Extensions/Core/IDialogEventService.cs
@@ -9,13 +9,13 @@ public interface IDialogEventService
     /// <summary>
     /// Subscribe to a dialog event
     /// </summary>
-    public void Subscribe<TEvent>(Func<TEvent, Task> handler) where TEvent : IDialogEvent;
+    public IDialogEventService Subscribe<TEvent>(Func<TEvent, Task> handler) where TEvent : IDialogEvent;
 
 
     /// <summary>
     /// Unsubscribe from a dialog event
     /// </summary>
-    public void Unsubscribe<TEvent>(Func<TEvent, Task> handler) where TEvent : IDialogEvent;
+    public IDialogEventService Unsubscribe<TEvent>(Func<TEvent, Task> handler) where TEvent : IDialogEvent;
 
     /// <summary>
     /// Publish a dialog event